Removed IO, Plotting, and Material related functions from rcs_functions module and put them into their own modules
