+++ src/open_rcs/rcs_functions.py
@@ -1,19 +1,42 @@
 from __future__ import annotations
 
 import cmath
-import importlib
 import math
-from collections.abc import Iterable
-from datetime import datetime
-from enum import Enum, IntEnum
 from pathlib import Path
-from typing import Any, Final, TypeAlias, cast
+from typing import Final, cast
 
-import matplotlib.pyplot as plt
 import numpy as np
 
-from .model_types import (
-    GeometryData,
+from .constants import (
+    RESULTS_DIR,
+    SPECIFIC_MATERIAL,
+    MaterialCode,
+    MaterialEntryIndex,
+    MaterialType,
+)
+from .materials import (
+    cart2spher,
+    compile_material_lookup_arrays,
+    convert_material_textlist_to_list,
+    get_entries_from_material_file,
+    get_material_properties_from_file,
+    get_reflection_coeff_from_material,
+    load_material_catalog,
+    refl_coeff,
+    reflection_coefficients,
+    rotation_transform_matrix,
+    save_list_in_file,
+    spher2cart,
+    spherical_global_to_local,
+)
+from .model_types import GeometryData
+from .plotting import (
+    final_plot,
+    generate_result_files,
+    plot_limits,
+    plot_parameters,
+    plot_triangle_model,
+    set_font_option,
 )
 from .stl_module import convert_stl
 
@@ -21,59 +44,7 @@ try:
     from numba import njit
 except ImportError:  # pragma: no cover - optional acceleration dependency
     njit = None
-yaml: Any
-try:
-    yaml = importlib.import_module("yaml")
-except ImportError:  # pragma: no cover - optional material library dependency
-    yaml = None
-
-
-class FontSize(IntEnum):
-    SMALL = 8
-    MEDIUM = 10
-    LARGE = 12
-
-
-class SphericalIndex(IntEnum):
-    THETA = 1
-
-
-class MaterialEntryIndex(IntEnum):
-    TYPE = 0
-    DESCRIPTION = 1
-    FIRST_LAYER = 2
-
-
-class MaterialType(str, Enum):
-    PEC = "PEC"
-    COMPOSITE = "Composite"
-    COMPOSITE_ON_PEC = "Composite Layer on PEC"
-    MULTI_LAYER = "Multiple Layers"
-    MULTI_LAYER_ON_PEC = "Multiple Layers on PEC"
-
-
-class MaterialCode(IntEnum):
-    PEC = 0
-    COMPOSITE = 1
-    COMPOSITE_ON_PEC = 2
-    MULTI_LAYER = 3
-    MULTI_LAYER_ON_PEC = 4
-
-
-MATERIAL_SPECIFIC: Final[float] = 1.0
-MATERIAL_TYPE_TO_CODE: Final[dict[str, int]] = {
-    MaterialType.PEC.value: int(MaterialCode.PEC),
-    MaterialType.COMPOSITE.value: int(MaterialCode.COMPOSITE),
-    MaterialType.COMPOSITE_ON_PEC.value: int(MaterialCode.COMPOSITE_ON_PEC),
-    MaterialType.MULTI_LAYER.value: int(MaterialCode.MULTI_LAYER),
-    MaterialType.MULTI_LAYER_ON_PEC.value: int(MaterialCode.MULTI_LAYER_ON_PEC),
-}
-RESULTS_DIR: Final[Path] = Path("./results")
 NUMBA_AVAILABLE: Final[bool] = njit is not None
-MaterialLayer: TypeAlias = list[float]
-MaterialEntry: TypeAlias = list[str | MaterialLayer]
-MaterialTable: TypeAlias = list[MaterialEntry]
-MaterialCatalog: TypeAlias = dict[str, MaterialEntry]
 _NUMBA_PLACEHOLDER_INT_ARRAY = np.zeros(1, dtype=np.int32)
 _NUMBA_PLACEHOLDER_FLOAT_ARRAY = np.zeros((1, 1), dtype=np.float64)
 
@@ -81,6 +52,32 @@ _MAT_CODE_PEC: Final[int] = int(MaterialCode.PEC)
 _MAT_CODE_COMPOSITE: Final[int] = int(MaterialCode.COMPOSITE)
 _MAT_CODE_COMPOSITE_ON_PEC: Final[int] = int(MaterialCode.COMPOSITE_ON_PEC)
 _MAT_CODE_MULTI_LAYER_ON_PEC: Final[int] = int(MaterialCode.MULTI_LAYER_ON_PEC)
+
+__all__ = [
+    "cart2spher",
+    "compile_material_lookup_arrays",
+    "convert_material_textlist_to_list",
+    "final_plot",
+    "generate_result_files",
+    "get_entries_from_material_file",
+    "get_material_properties_from_file",
+    "get_reflection_coeff_from_material",
+    "load_material_catalog",
+    "MaterialEntryIndex",
+    "MaterialType",
+    "plot_limits",
+    "plot_parameters",
+    "plot_triangle_model",
+    "refl_coeff",
+    "reflection_coefficients",
+    "RESULTS_DIR",
+    "rotation_transform_matrix",
+    "save_list_in_file",
+    "set_font_option",
+    "SPECIFIC_MATERIAL",
+    "spher2cart",
+    "spherical_global_to_local",
+]
 
 
 def get_polarization(incident_polarization: int) -> tuple[str, complex, complex]:
@@ -90,52 +87,6 @@ def get_polarization(incident_polarization: int) -> tuple[str, complex, complex]
     if incident_polarization == 1:
         return "TE-z", 0 + 0j, 1 + 0j
     raise ValueError("Invalid polarization value. Use 0 (TM-z) or 1 (TE-z).")
-
-
-def compile_material_lookup_arrays(
-    material_table: MaterialTable,
-) -> tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
-    """Convert heterogeneous material entries into dense numeric arrays for fast kernels."""
-    facet_count = len(material_table)
-    max_layers = max(
-        (len(entry) - MaterialEntryIndex.FIRST_LAYER for entry in material_table), default=0
-    )
-    max_layers = max(1, max_layers)
-
-    material_type_codes = np.zeros(facet_count, dtype=np.int32)
-    material_layer_count = np.zeros(facet_count, dtype=np.int32)
-    epsilon_r = np.zeros((facet_count, max_layers), dtype=np.float64)
-    loss_tangent = np.zeros((facet_count, max_layers), dtype=np.float64)
-    mu_r_real = np.zeros((facet_count, max_layers), dtype=np.float64)
-    mu_r_imag = np.zeros((facet_count, max_layers), dtype=np.float64)
-    thickness_m = np.zeros((facet_count, max_layers), dtype=np.float64)
-
-    for facet_index, entry in enumerate(material_table):
-        material_type = str(entry[MaterialEntryIndex.TYPE])
-        if material_type not in MATERIAL_TYPE_TO_CODE:
-            raise ValueError(f"Unsupported material type: {material_type}")
-        material_type_codes[facet_index] = MATERIAL_TYPE_TO_CODE[material_type]
-
-        layers = cast(list[list[float]], entry[MaterialEntryIndex.FIRST_LAYER :])
-        material_layer_count[facet_index] = len(layers)
-        for layer_index, layer in enumerate(layers):
-            if layer_index >= max_layers:
-                break
-            epsilon_r[facet_index, layer_index] = float(layer[0])
-            loss_tangent[facet_index, layer_index] = float(layer[1])
-            mu_r_real[facet_index, layer_index] = float(layer[2])
-            mu_r_imag[facet_index, layer_index] = float(layer[3])
-            thickness_m[facet_index, layer_index] = float(layer[4]) * 1e-3
-
-    return (
-        material_type_codes,
-        material_layer_count,
-        epsilon_r,
-        loss_tangent,
-        mu_r_real,
-        mu_r_imag,
-        thickness_m,
-    )
 
 
 def get_standard_deviation(
@@ -150,25 +101,6 @@ def get_standard_deviation(
     lt = 1e-5  # Taylor-series region
     nt = 5  # Number of terms in Taylor expansion
     return bk, cfac1, cfac2, rad, lt, nt
-
-
-def set_font_option(
-    font_size: int = int(FontSize.SMALL),
-    axes_title: int = int(FontSize.MEDIUM),
-    axes_label: int = int(FontSize.SMALL),
-    xtick_label: int = int(FontSize.SMALL),
-    ytick_label: int = int(FontSize.SMALL),
-    legend_size: int = int(FontSize.SMALL),
-    figure_title: int = int(FontSize.LARGE),
-) -> None:
-    """Configure global matplotlib font sizes."""
-    plt.rc("font", size=font_size)
-    plt.rc("axes", titlesize=axes_title)
-    plt.rc("axes", labelsize=axes_label)
-    plt.rc("xtick", labelsize=xtick_label)
-    plt.rc("ytick", labelsize=ytick_label)
-    plt.rc("legend", fontsize=legend_size)
-    plt.rc("figure", titlesize=figure_title)
 
 
 def extract_coordinates_data(rs_value: float) -> GeometryData:  # pragma: no cover
@@ -279,54 +211,6 @@ def calculate_r(x: np.ndarray, y: np.ndarray, z: np.ndarray, nverts: int) -> np.
     if nverts != len(x):
         raise ValueError("nverts must match the length of x/y/z arrays.")
     return np.column_stack((x, y, z)).astype(np.double, copy=False)
-
-
-def plot_triangle_model(
-    input_model, vind, x, y, z, xpts, ypts, zpts, nverts, ntria, node1, node2, node3, nfc
-):
-    fig = plt.figure(1)
-    fig.suptitle(f"Triangle Model of Target: {input_model}")
-    ax = fig.add_subplot(1, 1, 1, projection="3d")
-
-    for i in range(ntria):
-        X = [x[vind[i, 0] - 1], x[vind[i, 1] - 1], x[vind[i, 2] - 1], x[vind[i, 0] - 1]]
-        Y = [y[vind[i, 0] - 1], y[vind[i, 1] - 1], y[vind[i, 2] - 1], y[vind[i, 0] - 1]]
-        Z = [z[vind[i, 0] - 1], z[vind[i, 1] - 1], z[vind[i, 2] - 1], z[vind[i, 0] - 1]]
-        ax.plot(X, Y, Z)
-
-    ax.set_xlabel("x")
-    ax.set_ylabel("y")
-    ax.set_zlabel("z")
-    xmax = max(xpts)
-    xmin = min(xpts)
-    ymax = max(ypts)
-    ymin = min(ypts)
-    zmax = max(zpts)
-    zmin = min(zpts)
-
-    # Define the largest axis interval to preserve aspect ratio
-    x_range = xmax - xmin
-    y_range = ymax - ymin
-    z_range = zmax - zmin
-    max_range = max(x_range, y_range, z_range)
-
-    # Adjust each axis limits to keep the same scale
-    ax.set_xlim([xmin, xmin + max_range])
-    ax.set_ylim([ymin, ymin + max_range])
-    ax.set_zlim([zmin, zmin + max_range])
-
-    # Keep equal scaling across x, y, and z axes
-    ax.set_box_aspect([1, 1, 1])
-
-    # save plots
-    RESULTS_DIR.mkdir(parents=True, exist_ok=True)
-    now = datetime.now().strftime("%Y%m%d%H%M%S")
-    fig_name = str(RESULTS_DIR / f"temp_{now}.jpg")
-    extent = ax.get_window_extent().transformed(fig.dpi_scale_trans.inverted())
-    fig.savefig(fig_name, bbox_inches=extent)
-    plt.close()
-
-    return fig_name
 
 
 def precompute_rotation_matrices(
@@ -684,990 +568,6 @@ def taylor_g(n, w):
             go = g
             g = (exp_jw - m * go) / jw
     return g
-
-
-def _canonical_material_type(material_type: str) -> str:
-    normalized_type = material_type.strip().lower()
-    aliases = {
-        "pec": MaterialType.PEC.value,
-        "composite": MaterialType.COMPOSITE.value,
-        "composite layer on pec": MaterialType.COMPOSITE_ON_PEC.value,
-        "multiple layers": MaterialType.MULTI_LAYER.value,
-        "multiple layers on pec": MaterialType.MULTI_LAYER_ON_PEC.value,
-        # Legacy aliases accepted for smoother migration.
-        "composito": MaterialType.COMPOSITE.value,
-        "camada de composito em pec": MaterialType.COMPOSITE_ON_PEC.value,
-        "multiplas camadas": MaterialType.MULTI_LAYER.value,
-        "multiplas camadas em pec": MaterialType.MULTI_LAYER_ON_PEC.value,
-    }
-    if normalized_type not in aliases:
-        raise ValueError(f"Unsupported material type: {material_type}")
-    return aliases[normalized_type]
-
-
-def _parse_material_layer(layer_data: Any, context: str) -> MaterialLayer:
-    if isinstance(layer_data, dict):
-        raw_values = [
-            layer_data.get("epsilon_r"),
-            layer_data.get("loss_tangent"),
-            layer_data.get("mu_r_real"),
-            layer_data.get("mu_r_imag"),
-            layer_data.get("thickness_mm"),
-        ]
-    elif isinstance(layer_data, (list, tuple)) and len(layer_data) == 5:
-        raw_values = list(layer_data)
-    else:
-        raise ValueError(
-            f"{context}: each layer must be either a mapping with "
-            "'epsilon_r/loss_tangent/mu_r_real/mu_r_imag/thickness_mm' or a 5-value list."
-        )
-
-    parsed_values: MaterialLayer = []
-    for value in raw_values:
-        if value is None:
-            raise ValueError(f"{context}: all layer fields must be present.")
-        try:
-            parsed_values.append(float(cast(Any, value)))
-        except (TypeError, ValueError) as exc:
-            raise ValueError(f"{context}: all layer values must be numeric.") from exc
-    return parsed_values
-
-
-def _clone_material_entry(material_entry: MaterialEntry) -> MaterialEntry:
-    cloned_entry: MaterialEntry = [
-        str(material_entry[MaterialEntryIndex.TYPE]),
-        str(material_entry[MaterialEntryIndex.DESCRIPTION]),
-    ]
-    for layer in material_entry[MaterialEntryIndex.FIRST_LAYER :]:
-        cloned_entry.append(list(cast(MaterialLayer, layer)))
-    return cloned_entry
-
-
-def _material_entry_from_yaml(
-    material_definition: dict[str, Any], index: int
-) -> tuple[str, MaterialEntry]:
-    material_id = str(material_definition.get("id", "")).strip()
-    if not material_id:
-        raise ValueError(f"materials[{index}] is missing a non-empty 'id'.")
-    material_type = _canonical_material_type(str(material_definition.get("type", "")))
-    description = str(material_definition.get("description", material_id))
-    layers_raw = material_definition.get("layers", [])
-    if layers_raw is None:
-        layers_raw = []
-    if not isinstance(layers_raw, list):
-        raise ValueError(f"materials[{index}]: 'layers' must be a list.")
-
-    material_entry: MaterialEntry = [material_type, description]
-    if material_type != MaterialType.PEC.value and not layers_raw:
-        raise ValueError(f"materials[{index}] ({material_id}) requires at least one layer.")
-    for layer_index, layer_data in enumerate(layers_raw):
-        context = f"materials[{index}].layers[{layer_index}]"
-        material_entry.append(_parse_material_layer(layer_data, context))
-    return material_id, material_entry
-
-
-def _parse_facet_selector(selector: Any, ntria: int, context: str) -> set[int]:
-    selected_facets: set[int] = set()
-    if isinstance(selector, str):
-        if selector.strip().lower() != "all":
-            raise ValueError(f"{context}: unsupported selector string '{selector}'.")
-        return set(range(ntria))
-    if not isinstance(selector, list):
-        raise ValueError(f"{context}: selector must be 'all' or a list.")
-
-    for item in selector:
-        if isinstance(item, (list, tuple)) and len(item) == 2:
-            start = int(item[0])
-            stop = int(item[1])
-            if stop < start:
-                start, stop = stop, start
-            if start < 1 or stop > ntria:
-                raise ValueError(
-                    f"{context}: range [{start}, {stop}] is outside valid facet IDs 1..{ntria}."
-                )
-            selected_facets.update(range(start - 1, stop))
-        else:
-            facet_id = int(item)
-            if facet_id < 1 or facet_id > ntria:
-                raise ValueError(
-                    f"{context}: facet ID {facet_id} is outside valid range 1..{ntria}."
-                )
-            selected_facets.add(facet_id - 1)
-    return selected_facets
-
-
-def _parse_tags(tag_definitions: Any, ntria: int) -> dict[str, set[int]]:
-    if tag_definitions is None:
-        return {}
-    if not isinstance(tag_definitions, dict):
-        raise ValueError("'tags' must be a mapping from tag name to facet selectors.")
-    parsed_tags: dict[str, set[int]] = {}
-    for tag_name, selector in tag_definitions.items():
-        if not isinstance(tag_name, str) or not tag_name.strip():
-            raise ValueError("Tag names must be non-empty strings.")
-        parsed_tags[tag_name] = _parse_facet_selector(
-            selector,
-            ntria,
-            context=f"tags.{tag_name}",
-        )
-    return parsed_tags
-
-
-def _resolve_assignment_selector(
-    assignment: dict[str, Any],
-    ntria: int,
-    parsed_tags: dict[str, set[int]],
-    assignment_index: int,
-) -> set[int]:
-    selection: set[int] = set()
-    explicit_selector_used = False
-
-    if "facets" in assignment:
-        explicit_selector_used = True
-        selection.update(
-            _parse_facet_selector(
-                assignment["facets"],
-                ntria,
-                context=f"assignments[{assignment_index}].facets",
-            )
-        )
-    if "facet_indices" in assignment:
-        explicit_selector_used = True
-        selection.update(
-            _parse_facet_selector(
-                assignment["facet_indices"],
-                ntria,
-                context=f"assignments[{assignment_index}].facet_indices",
-            )
-        )
-    if "facet_ranges" in assignment:
-        explicit_selector_used = True
-        selection.update(
-            _parse_facet_selector(
-                assignment["facet_ranges"],
-                ntria,
-                context=f"assignments[{assignment_index}].facet_ranges",
-            )
-        )
-
-    for key in ("tags", "groups"):
-        if key in assignment:
-            explicit_selector_used = True
-            raw_names = assignment[key]
-            if not isinstance(raw_names, list):
-                raise ValueError(
-                    f"assignments[{assignment_index}].{key} must be a list of tag names."
-                )
-            for raw_name in raw_names:
-                tag_name = str(raw_name)
-                if tag_name not in parsed_tags:
-                    raise ValueError(
-                        f"assignments[{assignment_index}].{key} references "
-                        f"unknown tag '{tag_name}'."
-                    )
-                selection.update(parsed_tags[tag_name])
-
-    if not explicit_selector_used:
-        return set(range(ntria))
-    if not selection:
-        raise ValueError(f"assignments[{assignment_index}] does not select any facets.")
-    return selection
-
-
-def _material_table_from_yaml_document(document: dict[str, Any], ntria: int) -> MaterialTable:
-    materials_raw = document.get("materials")
-    if not isinstance(materials_raw, list) or not materials_raw:
-        raise ValueError("YAML material file must define a non-empty 'materials' list.")
-
-    material_catalog: MaterialCatalog = {}
-    for material_index, raw_material in enumerate(materials_raw):
-        if not isinstance(raw_material, dict):
-            raise ValueError(f"materials[{material_index}] must be a mapping.")
-        material_id, material_entry = _material_entry_from_yaml(raw_material, material_index)
-        if material_id in material_catalog:
-            raise ValueError(f"Duplicate material id '{material_id}' in YAML file.")
-        material_catalog[material_id] = material_entry
-
-    default_material = document.get("default_material")
-    if default_material is None and isinstance(document.get("defaults"), dict):
-        default_material = cast(dict[str, Any], document["defaults"]).get("material")
-    default_material_id = str(default_material).strip() if default_material is not None else ""
-    if default_material_id and default_material_id not in material_catalog:
-        raise ValueError(f"default_material '{default_material_id}' is not defined in materials.")
-
-    parsed_tags = _parse_tags(document.get("tags"), ntria)
-    assignments_raw = document.get("assignments", [])
-    if assignments_raw is None:
-        assignments_raw = []
-    if not isinstance(assignments_raw, list):
-        raise ValueError("'assignments' must be a list when provided.")
-
-    facet_material_ids: list[str | None] = [None] * ntria
-    if default_material_id:
-        for facet_index in range(ntria):
-            facet_material_ids[facet_index] = default_material_id
-
-    for assignment_index, raw_assignment in enumerate(assignments_raw):
-        if not isinstance(raw_assignment, dict):
-            raise ValueError(f"assignments[{assignment_index}] must be a mapping.")
-        material_id = str(raw_assignment.get("material", "")).strip()
-        if not material_id:
-            raise ValueError(f"assignments[{assignment_index}] is missing 'material'.")
-        if material_id not in material_catalog:
-            raise ValueError(
-                f"assignments[{assignment_index}] references unknown material id '{material_id}'."
-            )
-        selected_facets = _resolve_assignment_selector(
-            raw_assignment,
-            ntria,
-            parsed_tags,
-            assignment_index,
-        )
-        for facet_index in selected_facets:
-            facet_material_ids[facet_index] = material_id
-
-    missing_facets = [
-        index + 1
-        for index, assigned_material_id in enumerate(facet_material_ids)
-        if assigned_material_id is None
-    ]
-    if missing_facets:
-        preview = ", ".join(str(value) for value in missing_facets[:10])
-        suffix = "..." if len(missing_facets) > 10 else ""
-        raise ValueError(
-            f"YAML material assignments are incomplete. Missing facet IDs: {preview}{suffix}"
-        )
-
-    material_table: MaterialTable = []
-    for assigned_material_id in facet_material_ids:
-        assert assigned_material_id is not None
-        material_table.append(_clone_material_entry(material_catalog[assigned_material_id]))
-    return material_table
-
-
-def load_material_catalog(material_path: str | Path) -> MaterialCatalog:
-    """Load and validate the material catalog section from a YAML .rcsmat file."""
-    if yaml is None:  # pragma: no cover - runtime dependency guard
-        raise ImportError("YAML material files require PyYAML. Install with: pip install pyyaml")
-
-    with Path(material_path).open("r", encoding="utf-8") as stream:
-        document = yaml.safe_load(stream)
-    if not isinstance(document, dict):
-        raise ValueError("YAML material file root must be a mapping.")
-
-    materials_raw = document.get("materials")
-    if not isinstance(materials_raw, list) or not materials_raw:
-        raise ValueError("YAML material file must define a non-empty 'materials' list.")
-
-    material_catalog: MaterialCatalog = {}
-    for material_index, raw_material in enumerate(materials_raw):
-        if not isinstance(raw_material, dict):
-            raise ValueError(f"materials[{material_index}] must be a mapping.")
-        material_id, material_entry = _material_entry_from_yaml(raw_material, material_index)
-        if material_id in material_catalog:
-            raise ValueError(f"Duplicate material id '{material_id}' in YAML file.")
-        material_catalog[material_id] = _clone_material_entry(material_entry)
-    return material_catalog
-
-
-def _load_material_table_from_yaml(material_path: str | Path, ntria: int) -> MaterialTable:
-    if yaml is None:  # pragma: no cover - runtime dependency guard
-        raise ImportError("YAML material files require PyYAML. Install with: pip install pyyaml")
-
-    with Path(material_path).open("r", encoding="utf-8") as stream:
-        document = yaml.safe_load(stream)
-    if not isinstance(document, dict):
-        raise ValueError("YAML material file root must be a mapping.")
-    return _material_table_from_yaml_document(document, ntria)
-
-
-def save_list_in_file(material_rows: list, output_file: str) -> None:
-    serialized_rows = []
-    for row in material_rows:
-        entry_str = (
-            str(row[MaterialEntryIndex.TYPE]) + "," + str(row[MaterialEntryIndex.DESCRIPTION])
-        )
-
-        for layer in row[MaterialEntryIndex.FIRST_LAYER :]:
-            for i in range(len(layer)):
-                entry_str = entry_str + "," + str(layer[i])
-
-        serialized_rows.append(entry_str)
-
-    with open(output_file, "w") as file:
-        for row in serialized_rows:
-            file.write(row + "\n")
-
-
-def get_entries_from_material_file(ntria: int, matrlpath: str) -> MaterialTable:
-    material_path = Path(matrlpath)
-    try:
-        if material_path.suffix.lower() in {".rcsmat", ".yaml", ".yml"}:
-            matrl = _load_material_table_from_yaml(material_path, ntria)
-        else:
-            with material_path.open("r", encoding="utf-8") as file:
-                matrl = get_material_properties_from_file(file)
-    except FileNotFoundError as exc:
-        raise FileNotFoundError(f"Material file not found: {material_path}") from exc
-    except Exception as exc:
-        raise ValueError(f"Invalid material file: {material_path}") from exc
-
-    if len(matrl) != ntria:
-        raise ValueError("Number of material entries does not match the number of facets.")
-
-    return matrl
-
-
-def get_material_properties_from_file(rows: Iterable[str]) -> MaterialTable:
-    material_text_rows = list(rows)
-    return convert_material_textlist_to_list(material_text_rows)
-
-
-def convert_material_textlist_to_list(text_rows: Iterable[str]) -> MaterialTable:
-    material_table: MaterialTable = []
-    for row in text_rows:
-        entries = row.strip("\n").split(",")
-        if len(entries) < 2:
-            continue
-        formatted_entries: MaterialEntry = [
-            _canonical_material_type(entries[MaterialEntryIndex.TYPE]),
-            entries[MaterialEntryIndex.DESCRIPTION],
-        ]
-        layer_values: MaterialLayer = []
-
-        for index, entry in enumerate(entries[MaterialEntryIndex.FIRST_LAYER :]):
-            layer_values.append(float(entry))
-            if (index + 1) % 5 == 0:
-                formatted_entries.append(layer_values)
-                layer_values = []
-
-        material_table.append(formatted_entries)
-    return material_table
-
-
-def rotation_transform_matrix(alpha, beta):
-    T1 = np.array(
-        [[np.cos(alpha), np.sin(alpha), 0], [-np.sin(alpha), np.cos(alpha), 0], [0, 0, 1]]
-    )
-    T2 = np.array([[np.cos(beta), 0, -np.sin(beta)], [0, 1, 0], [np.sin(beta), 0, np.cos(beta)]])
-    return np.dot(T2, T1)
-
-
-def refl_coeff(er1, mr1, er2, mr2, thetai):
-    m0 = 4 * np.pi * 1e-7  # vacuum permeability
-    e0 = 8.854e-12  # vacuum permittivity
-
-    TIR = 0
-    sinthetat = np.sin(thetai) * np.sqrt(
-        np.real(er1) * np.real(mr1) / (np.real(er2) * np.real(mr2))
-    )
-
-    # Total internal reflection (TIR) check
-    if sinthetat > 1:
-        TIR = 1
-        thetat = np.pi / 2  # critical angle
-    else:
-        thetat = np.arcsin(sinthetat)
-
-    # Compute n1 and n2 (refractive indices)
-    n1 = np.sqrt(mr1 * m0 / (er1 * e0))
-    n2 = np.sqrt(mr2 * m0 / (er2 * e0))
-
-    # Compute gammaperp and gammapar
-    gammaperp = (n2 * np.cos(thetai) - n1 * np.cos(thetat)) / (
-        n2 * np.cos(thetai) + n1 * np.cos(thetat)
-    )
-    gammapar = (n2 * np.cos(thetat) - n1 * np.cos(thetai)) / (
-        n2 * np.cos(thetat) + n1 * np.cos(thetai)
-    )
-
-    return gammapar, gammaperp, thetat, TIR
-
-
-def _resolve_local_incidence_theta(
-    thri: float,
-    phrii: float,
-    alpha: float,
-    beta: float,
-    local_theta: float | None,
-) -> float:
-    """Resolve local incidence angle, using precomputed local value when available."""
-    if local_theta is not None:
-        return float(local_theta)
-    transform_global_to_local = rotation_transform_matrix(alpha, beta)
-    spherical_vector = spherical_global_to_local(
-        np.array([1.0, thri, phrii], dtype=float),
-        transform_global_to_local,
-    )
-    return float(spherical_vector[SphericalIndex.THETA])
-
-
-def _mat2_mul(
-    a00: complex,
-    a01: complex,
-    a10: complex,
-    a11: complex,
-    b00: complex,
-    b01: complex,
-    b10: complex,
-    b11: complex,
-) -> tuple[complex, complex, complex, complex]:
-    """Multiply two 2x2 complex matrices represented as scalar entries."""
-    return (
-        a00 * b00 + a01 * b10,
-        a00 * b01 + a01 * b11,
-        a10 * b00 + a11 * b10,
-        a10 * b01 + a11 * b11,
-    )
-
-
-def _layer_properties(layer: list[float]) -> tuple[complex, complex, float]:
-    epsilon_r = layer[0]
-    loss_tangent = layer[1]
-    mu_r_real = layer[2]
-    mu_r_imag = layer[3]
-    thickness_m = layer[4] * 1e-3
-    epsilon_complex = epsilon_r - 1j * loss_tangent * epsilon_r
-    mu_complex = mu_r_real - 1j * mu_r_imag
-    return epsilon_complex, mu_complex, thickness_m
-
-
-def _apply_transfer_matrix(
-    m00: complex,
-    m01: complex,
-    m10: complex,
-    m11: complex,
-    gamma: complex,
-    phase: complex,
-) -> tuple[complex, complex, complex, complex]:
-    exp_phase = cmath.exp(1j * phase)
-    exp_neg_phase = cmath.exp(-1j * phase)
-    t00 = exp_phase
-    t01 = gamma * exp_neg_phase
-    t10 = gamma * exp_phase
-    t11 = exp_neg_phase
-    return _mat2_mul(m00, m01, m10, m11, t00, t01, t10, t11)
-
-
-def spher2cart(spherical_vector: np.ndarray) -> np.ndarray:
-    radius = float(spherical_vector[0])
-    theta = float(spherical_vector[1])
-    phi = float(spherical_vector[2])
-    x = radius * np.sin(theta) * np.cos(phi)
-    y = radius * np.sin(theta) * np.sin(phi)
-    z = radius * np.cos(theta)
-
-    return np.array([x, y, z])
-
-
-def cart2spher(cart_vector: np.ndarray) -> np.ndarray:
-    x = float(cart_vector[0])
-    y = float(cart_vector[1])
-    z = float(cart_vector[2])
-    radius = np.sqrt(x**2 + y**2 + z**2)
-    theta = math.atan2(np.sqrt(x**2 + y**2), z)
-    phi = math.atan2(y, x)
-    return np.array([radius, theta, phi], dtype=float)
-
-
-def spherical_global_to_local(
-    spherical_vector: np.ndarray,
-    transform_matrix_global_to_local: np.ndarray,
-) -> np.ndarray:
-    cartesian_vector = spher2cart(spherical_vector)
-    cartesian_vector = np.dot(transform_matrix_global_to_local, cartesian_vector)
-    return cart2spher(cartesian_vector)
-
-
-def refl_coeff_composite(
-    thri: float,
-    phrii: float,
-    alpha: float,
-    beta: float,
-    freq: float,
-    matrlLine: list,
-    local_theta: float | None = None,
-) -> tuple[complex, complex]:
-    layer = cast(list[float], matrlLine[MaterialEntryIndex.FIRST_LAYER])
-    local_incidence_theta = _resolve_local_incidence_theta(thri, phrii, alpha, beta, local_theta)
-
-    epsilon_complex, mu_complex, thickness_m = _layer_properties(layer)
-    gamma_parallel_1, gamma_perpendicular_1, _theta_t, _tir = refl_coeff(
-        1,
-        1,
-        epsilon_complex,
-        mu_complex,
-        local_incidence_theta,
-    )
-    gamma_parallel_2 = -gamma_parallel_1
-    gamma_perpendicular_2 = -gamma_perpendicular_1
-
-    wave_speed = 3e8 / math.sqrt(np.real(epsilon_complex) * np.real(mu_complex))
-    phase = 2.0 * math.pi * thickness_m / (wave_speed / freq)
-
-    m00p, m01p, m10p, m11p = _apply_transfer_matrix(
-        1.0 + 0j,
-        0.0 + 0j,
-        0.0 + 0j,
-        1.0 + 0j,
-        complex(gamma_parallel_1),
-        phase,
-    )
-    m00s, m01s, m10s, m11s = _apply_transfer_matrix(
-        1.0 + 0j,
-        0.0 + 0j,
-        0.0 + 0j,
-        1.0 + 0j,
-        complex(gamma_perpendicular_1),
-        phase,
-    )
-    m00p, m01p, m10p, m11p = _mat2_mul(
-        m00p,
-        m01p,
-        m10p,
-        m11p,
-        1.0 + 0j,
-        complex(gamma_parallel_2),
-        complex(gamma_parallel_2),
-        1.0 + 0j,
-    )
-    m00s, m01s, m10s, m11s = _mat2_mul(
-        m00s,
-        m01s,
-        m10s,
-        m11s,
-        1.0 + 0j,
-        complex(gamma_perpendicular_2),
-        complex(gamma_perpendicular_2),
-        1.0 + 0j,
-    )
-    reflection_parallel = m10p / m00p
-    reflection_perpendicular = m10s / m00s
-    return reflection_perpendicular, reflection_parallel
-
-
-def refl_coeff_composite_layer_on_pec(
-    thri: float,
-    phrii: float,
-    alpha: float,
-    beta: float,
-    freq: float,
-    matrlLine: list,
-    local_theta: float | None = None,
-) -> tuple[complex, complex]:
-    layers = cast(list[list[float]], matrlLine[MaterialEntryIndex.FIRST_LAYER :])
-    local_incidence_theta = _resolve_local_incidence_theta(thri, phrii, alpha, beta, local_theta)
-    sine_incidence = math.sin(local_incidence_theta)
-    cosine_incidence = math.cos(local_incidence_theta)
-
-    wave = 3e8 / freq
-    beta_0 = 2.0 * math.pi / wave
-    z_0 = 1.0 + 0j
-
-    w00p, w01p, w10p, w11p = 1.0 + 0j, 0.0 + 0j, 0.0 + 0j, 1.0 + 0j
-    w00s, w01s, w10s, w11s = 1.0 + 0j, 0.0 + 0j, 0.0 + 0j, 1.0 + 0j
-    previous_z_parallel: complex | None = None
-    previous_z_perpendicular: complex | None = None
-
-    for layer_index, layer in enumerate(layers):
-        epsilon_complex, mu_complex, thickness_m = _layer_properties(layer)
-        impedance_ratio = epsilon_complex / mu_complex
-        root_term = cmath.sqrt(impedance_ratio - sine_incidence**2)
-        z_parallel = root_term / (impedance_ratio * cosine_incidence)
-        z_perpendicular = cosine_incidence / root_term
-
-        if layer_index == 0:
-            gamma_parallel = (z_parallel - z_0) / (z_parallel + z_0)
-            gamma_perpendicular = (z_perpendicular - z_0) / (z_perpendicular + z_0)
-        else:
-            assert previous_z_parallel is not None
-            assert previous_z_perpendicular is not None
-            gamma_parallel = (z_parallel - previous_z_parallel) / (z_parallel + previous_z_parallel)
-            gamma_perpendicular = (z_perpendicular - previous_z_perpendicular) / (
-                z_perpendicular + previous_z_perpendicular
-            )
-
-        previous_z_parallel = z_parallel
-        previous_z_perpendicular = z_perpendicular
-        tau_parallel = 1.0 + gamma_parallel
-        tau_perpendicular = 1.0 + gamma_perpendicular
-        phase = beta_0 * thickness_m * cmath.sqrt(epsilon_complex * mu_complex - sine_incidence**2)
-
-        w00p, w01p, w10p, w11p = _apply_transfer_matrix(
-            w00p,
-            w01p,
-            w10p,
-            w11p,
-            gamma_parallel,
-            phase,
-        )
-        w00s, w01s, w10s, w11s = _apply_transfer_matrix(
-            w00s,
-            w01s,
-            w10s,
-            w11s,
-            gamma_perpendicular,
-            phase,
-        )
-        inverse_tau_parallel = 1.0 / tau_parallel
-        inverse_tau_perpendicular = 1.0 / tau_perpendicular
-        w00p *= inverse_tau_parallel
-        w01p *= inverse_tau_parallel
-        w10p *= inverse_tau_parallel
-        w11p *= inverse_tau_parallel
-        w00s *= inverse_tau_perpendicular
-        w01s *= inverse_tau_perpendicular
-        w10s *= inverse_tau_perpendicular
-        w11s *= inverse_tau_perpendicular
-
-    # Right-multiply by PEC matrix [[1, 0], [-1, 0]] and take ratio [1,0]/[0,0].
-    reflection_parallel = (w10p - w11p) / (w00p - w01p)
-    reflection_perpendicular = (w10s - w11s) / (w00s - w01s)
-    return reflection_perpendicular, reflection_parallel
-
-
-def refl_coeff_multi_layers(
-    thri: float,
-    phrii: float,
-    alpha: float,
-    beta: float,
-    freq: float,
-    matrlLine: list,
-    local_theta: float | None = None,
-) -> tuple[complex, complex]:
-    layers = cast(list[list[float]], matrlLine[MaterialEntryIndex.FIRST_LAYER :])
-    local_incidence_theta = _resolve_local_incidence_theta(thri, phrii, alpha, beta, local_theta)
-
-    m00p, m01p, m10p, m11p = 1.0 + 0j, 0.0 + 0j, 0.0 + 0j, 1.0 + 0j
-    m00s, m01s, m10s, m11s = 1.0 + 0j, 0.0 + 0j, 0.0 + 0j, 1.0 + 0j
-
-    previous_epsilon = 1.0 + 0j
-    previous_mu = 1.0 + 0j
-    previous_theta = local_incidence_theta
-    last_phase = 0.0 + 0j
-
-    for layer_index, layer in enumerate(layers):
-        epsilon_complex, mu_complex, thickness_m = _layer_properties(layer)
-        if layer_index == 0:
-            gamma_parallel, gamma_perpendicular, theta_transmitted, _tir = refl_coeff(
-                1.0 + 0j,
-                1.0 + 0j,
-                epsilon_complex,
-                mu_complex,
-                local_incidence_theta,
-            )
-        else:
-            gamma_parallel, gamma_perpendicular, theta_transmitted, _tir = refl_coeff(
-                previous_epsilon,
-                previous_mu,
-                epsilon_complex,
-                mu_complex,
-                previous_theta,
-            )
-
-        wave_speed = 3e8 / math.sqrt(np.real(epsilon_complex) * np.real(mu_complex))
-        last_phase = 2.0 * math.pi * thickness_m / (wave_speed / freq)
-        m00p, m01p, m10p, m11p = _apply_transfer_matrix(
-            m00p,
-            m01p,
-            m10p,
-            m11p,
-            complex(gamma_parallel),
-            last_phase,
-        )
-        m00s, m01s, m10s, m11s = _apply_transfer_matrix(
-            m00s,
-            m01s,
-            m10s,
-            m11s,
-            complex(gamma_perpendicular),
-            last_phase,
-        )
-        previous_epsilon = epsilon_complex
-        previous_mu = mu_complex
-        previous_theta = float(theta_transmitted)
-
-    gamma_parallel_exit, gamma_perpendicular_exit, _theta_exit, _tir = refl_coeff(
-        previous_epsilon,
-        previous_mu,
-        1.0 + 0j,
-        1.0 + 0j,
-        previous_theta,
-    )
-    m00p, m01p, m10p, m11p = _apply_transfer_matrix(
-        m00p,
-        m01p,
-        m10p,
-        m11p,
-        complex(gamma_parallel_exit),
-        last_phase,
-    )
-    m00s, m01s, m10s, m11s = _apply_transfer_matrix(
-        m00s,
-        m01s,
-        m10s,
-        m11s,
-        complex(gamma_perpendicular_exit),
-        last_phase,
-    )
-    reflection_parallel = m10p / m00p
-    reflection_perpendicular = m10s / m00s
-    return reflection_perpendicular, reflection_parallel
-
-
-def refl_coeff_multi_layers_on_pec(
-    thri: float,
-    phrii: float,
-    alpha: float,
-    beta: float,
-    freq: float,
-    matrlLine: list,
-    local_theta: float | None = None,
-) -> tuple[complex, complex]:
-    return refl_coeff_composite_layer_on_pec(
-        thri=thri,
-        phrii=phrii,
-        alpha=alpha,
-        beta=beta,
-        freq=freq,
-        matrlLine=matrlLine,
-        local_theta=local_theta,
-    )
-
-
-def get_reflection_coeff_from_material(
-    thri: float,
-    phrii: float,
-    alpha: float,
-    beta: float,
-    freq: float,
-    matrlLine: list,
-    local_theta: float | None = None,
-) -> tuple[complex, complex]:
-    reflection_perpendicular: complex = 0.0 + 0.0j
-    reflection_parallel: complex = 0.0 + 0.0j
-
-    if matrlLine[MaterialEntryIndex.TYPE] == MaterialType.PEC.value:
-        reflection_perpendicular = -1.0 + 0.0j
-        reflection_parallel = -1.0 + 0.0j
-
-    elif matrlLine[MaterialEntryIndex.TYPE] == MaterialType.COMPOSITE.value:
-        reflection_perpendicular, reflection_parallel = refl_coeff_composite(
-            thri,
-            phrii,
-            alpha,
-            beta,
-            freq,
-            matrlLine,
-            local_theta=local_theta,
-        )
-
-    elif matrlLine[MaterialEntryIndex.TYPE] == MaterialType.COMPOSITE_ON_PEC.value:
-        reflection_perpendicular, reflection_parallel = refl_coeff_composite_layer_on_pec(
-            thri,
-            phrii,
-            alpha,
-            beta,
-            freq,
-            matrlLine,
-            local_theta=local_theta,
-        )
-
-    elif matrlLine[MaterialEntryIndex.TYPE] == MaterialType.MULTI_LAYER.value:
-        reflection_perpendicular, reflection_parallel = refl_coeff_multi_layers(
-            thri,
-            phrii,
-            alpha,
-            beta,
-            freq,
-            matrlLine,
-            local_theta=local_theta,
-        )
-
-    elif matrlLine[MaterialEntryIndex.TYPE] == MaterialType.MULTI_LAYER_ON_PEC.value:
-        reflection_perpendicular, reflection_parallel = refl_coeff_multi_layers_on_pec(
-            thri,
-            phrii,
-            alpha,
-            beta,
-            freq,
-            matrlLine,
-            local_theta=local_theta,
-        )
-    else:
-        material_type = str(matrlLine[MaterialEntryIndex.TYPE])
-        raise ValueError(f"Unsupported material type: {material_type}")
-
-    return reflection_perpendicular, reflection_parallel
-
-
-def reflection_coefficients(
-    rs: float,
-    index: int,
-    th2: float,
-    thri: float,
-    phrii: float,
-    alpha: float,
-    beta: float,
-    freq: float,
-    matrl: list,
-    local_cos_theta: float | None = None,
-) -> tuple[complex, complex]:
-    reflection_perpendicular: complex = 0.0 + 0.0j
-    reflection_parallel: complex = 0.0 + 0.0j
-
-    if rs == MATERIAL_SPECIFIC:
-        reflection_perpendicular, reflection_parallel = get_reflection_coeff_from_material(
-            thri,
-            phrii,
-            alpha,
-            beta,
-            freq,
-            matrl[index],
-            local_theta=th2,
-        )
-    else:
-        cosine_local_theta = math.cos(th2) if local_cos_theta is None else local_cos_theta
-        reflection_perpendicular = complex(
-            -1 / (2 * rs * cosine_local_theta + 1)
-        )  # local TE polarization
-        reflection_parallel = 0.0 + 0.0j  # local TM polarization
-        if (2 * rs + cosine_local_theta) != 0:
-            reflection_parallel = complex(-cosine_local_theta / (2 * rs + cosine_local_theta))
-
-    return reflection_perpendicular, reflection_parallel
-
-
-def final_plot(
-    phi_sample_count: int,
-    theta_sample_count: int,
-    phi_grid_deg: np.ndarray,
-    wavelength_m: float,
-    theta_grid_deg: np.ndarray,
-    rcs_min_db: float,
-    rcs_max_db: float,
-    rcs_theta_db: np.ndarray,
-    rcs_phi_db: np.ndarray,
-    direction_cosine_u_grid: np.ndarray,
-    direction_cosine_v_grid: np.ndarray,
-    timestamp: str,
-    input_model: str,
-    mode: str,
-) -> str:
-    if phi_sample_count == 1:
-        plt.figure(1)
-        plt.suptitle(f"RCS Simulation IR Signature - {mode}")
-        plt.title(
-            f"target: {input_model}   solid: theta     dashed: phi     "
-            f"phi= {phi_grid_deg[0][0]}    wave (m): {round(wavelength_m, 6)}"
-        )
-        plt.xlabel(f"{mode} Angle, theta (deg)")
-        plt.ylabel("RCS (dBsm)")
-        plt.axis(
-            (
-                float(np.min(theta_grid_deg)),
-                float(np.max(theta_grid_deg)),
-                float(rcs_min_db),
-                float(rcs_max_db),
-            )
-        )
-        plt.plot(theta_grid_deg[0], rcs_theta_db[0])
-        plt.plot(theta_grid_deg[0], rcs_phi_db[0], linewidth=2, linestyle="dashed")
-        plt.grid(True)
-
-    if theta_sample_count == 1:
-        plt.figure(1)
-        plt.suptitle(f"RCS Simulation IR Signature - {mode}")
-        plt.title(
-            f"target: {input_model}   solid: theta     dashed: phi     "
-            f"theta= {theta_grid_deg[0][0]}    wave (m): {round(wavelength_m, 6)}"
-        )
-        plt.xlabel(f"{mode} Angle, phi (deg)")
-        plt.ylabel("RCS (dBsm)")
-        plt.axis(
-            (
-                float(np.min(phi_grid_deg)),
-                float(np.max(phi_grid_deg)),
-                float(rcs_min_db),
-                float(rcs_max_db),
-            )
-        )
-        plt.plot(phi_grid_deg, rcs_theta_db)
-        plt.plot(phi_grid_deg, rcs_phi_db, linewidth=2, linestyle="dashed")
-        plt.grid(True)
-
-    if phi_sample_count > 1 and theta_sample_count > 1:
-        contour_levels = [-20, 0]
-        fig = plt.figure(1)
-        fig.suptitle(f"RCS Simulation IR Signature - {mode}")
-
-        ax = fig.add_subplot(2, 3, 2)
-        if mode == "Monostatic":
-            cp = ax.contour(direction_cosine_u_grid, direction_cosine_v_grid, rcs_theta_db)
-        elif mode == "Bistatic":
-            cp = ax.contour(
-                direction_cosine_u_grid, direction_cosine_v_grid, rcs_theta_db, contour_levels
-            )
-        ax.set_title("RCS-theta")
-        ax.set_xlabel("U")
-        ax.set_ylabel("V")
-        ax.axis("square")
-        cbar = fig.colorbar(cp)
-        cbar.set_label("RCS (dBsm)")
-
-        bx = fig.add_subplot(2, 3, 5)
-        if mode == "Monostatic":
-            cp = bx.contour(direction_cosine_u_grid, direction_cosine_v_grid, rcs_phi_db)
-        elif mode == "Bistatic":
-            cp = bx.contour(
-                direction_cosine_u_grid, direction_cosine_v_grid, rcs_phi_db, contour_levels
-            )
-        bx.set_title("RCS-phi")
-        bx.set_xlabel("U")
-        bx.set_ylabel("V")
-        bx.axis("square")
-        cbar = fig.colorbar(cp)
-        cbar.set_label("RCS (dBsm)")
-
-        fig.subplots_adjust(wspace=0)
-
-    RESULTS_DIR.mkdir(parents=True, exist_ok=True)
-    plot_name = str(RESULTS_DIR / f"temp_{timestamp}.png")
-    plt.savefig(plot_name)
-    plt.close()
-    return plot_name
-
-
-def generate_result_files(
-    theta_grid_deg: np.ndarray,
-    rcs_theta_db: np.ndarray,
-    phi_grid_deg: np.ndarray,
-    rcs_phi_db: np.ndarray,
-    parameter_text: str,
-    phi_sample_count: int,
-):
-    RESULTS_DIR.mkdir(parents=True, exist_ok=True)
-    now = datetime.now().strftime("%Y%m%d%H%M%S")
-    file_name = str(RESULTS_DIR / f"temp_{now}.dat")
-
-    with open(file_name, "w", encoding="utf-8") as result_file:
-        result_file.write(f"RCS SIMULATOR RESULTS {now}\n")
-        result_file.write("\nSimulation Parameters:\n" + parameter_text)
-        result_file.write("\nSimulation Results IR Signature:")
-        result_file.write("\nTheta (deg):\n")
-        for i1 in range(phi_sample_count):
-            result_file.write(f"{theta_grid_deg[i1]}\n")
-        result_file.write("\nRCS Theta (dBsm):\n")
-        for i1 in range(phi_sample_count):
-            result_file.write(f"{rcs_theta_db[i1]}\n")
-        result_file.write("\nPhi (deg):\n")
-        for i1 in range(phi_sample_count):
-            result_file.write(f"{phi_grid_deg[i1]}\n")
-        result_file.write("\nRCS Phi (dBsm):\n")
-        for i1 in range(phi_sample_count):
-            result_file.write(f"{rcs_phi_db[i1]}\n")
-
-    return now, file_name
 
 
 def calculate_ic(
@@ -3039,36 +1939,6 @@ def accumulate_bistatic_sample_numba_material(
     )
 
 
-def plot_parameters(
-    mode: str,
-    frequency_hz: float,
-    wavelength_m: float,
-    correlation_distance_m: float,
-    standard_deviation_m: float,
-    polarization_label: str,
-    triangle_count: int,
-    phi_start_deg: float,
-    phi_stop_deg: float,
-    phi_step_deg: float,
-    theta_start_deg: float,
-    theta_stop_deg: float,
-    theta_step_deg: float,
-):
-    param = f"    Mode: {mode}\n\
-    Radar Frequency (GHz): {frequency_hz / 1e9}\n\
-    Wavelength (m): {wavelength_m}\n\
-    Correlation distance (m): {correlation_distance_m}\n\
-    Standard Deviation (m): {standard_deviation_m}\n\
-    Incident wave polarization: {polarization_label}\n\
-    Start phi angle (degrees): {phi_start_deg}\n\
-    Stop phi angle (degrees): {phi_stop_deg}\n\
-    Phi increment step (degrees): {phi_step_deg}\n\
-    Start theta angle (degrees): {theta_start_deg}\n\
-    Stop theta angle (degrees): {theta_stop_deg}\n\
-    Phi increment step (degrees): {theta_step_deg}\n"
-    return param
-
-
 def calculate_fields(
     triangle_area: float,
     roughness_factor_secondary: float,
@@ -3178,13 +2048,6 @@ def calculate_sth_sph(
     )
     rcs_theta_db[phi_index, theta_index] = 10 * np.log10(np.maximum(theta_power, floor_power))
     rcs_phi_db[phi_index, theta_index] = 10 * np.log10(np.maximum(phi_power, floor_power))
-
-
-def plot_limits(rcs_theta_db: np.ndarray, rcs_phi_db: np.ndarray):
-    rcs_max_db = max(np.max(rcs_theta_db), np.max(rcs_phi_db))
-    rounded_max_db = (np.floor(rcs_max_db / 5) + 1) * 5
-    rcs_min_db = min(np.min(rcs_theta_db), np.min(rcs_phi_db))
-    return rounded_max_db, rcs_min_db
 
 
 def product_vector(
